requirements.txt
fastapi
uvicorn
pywinauto; sys_platform == "win32"
pywinpty; sys_platform == "win32"
websockets
